+++ packages/utils-chuni/src/index.ts
@@ -1,3 +1,4 @@
+export * from "./decimal";
 export * from "./rank";
 export * from "./rarity";
 export * from "./rating";